Unfiltering optimizations
Optimization changes:
* Don't 'shift' the image data as it is being unfiltered
The prior code would 'compact' the data when processing each and every pixel by shifting it left to overwrite the byte indicating the filter to use.
Instead, simplify a,b,c pixel calculations by leaving these intact.
To prevent major refactors, we just 'compact' the rows after we finish processing them in `unfilter`
* Change filter functions to work on whole rows at a time rather than individual pixels.
Each row has a filter function set on it, so this saves us having to ecase on the filter per-pixel.
* Make first-column and first-row optimizations
In the `sub` filter, we can skip the very first pixel, which eliminates a branch per pixel in checking if we're on the first column
In the `up` filter we can likewise skip the first row entirely
In the `paeth` filter we can just sub out work to `sub` on the first row, and we can take care of the first column explicitly otherwise to avoid branching for it as well.

With these changes, On My Machine it yields about a 20% performance increase on average.